update handling of device mapping
